pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "enveloper"
version = "0.1.14"
description = "Manage .env secrets via system keychain with cloud store plugins"
readme = "README.md"
requires-python = ">=3.10"
license = "AGPL-3.0-or-later"
authors = [{ name = "Ramin Firoozye" }]
keywords = ["secrets", "keychain", "env", "dotenv", "ssm", "github-secrets"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "keyring>=25.0",
    "tomli>=2.0; python_version < '3.11'",
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
cli = []  # CLI is included in base; [cli] for use with [cli,sdk]
sdk = []  # SDK (load_dotenv / dotenv_values); use [sdk] or [cli,sdk] to add SDK
aws = ["boto3>=1.35"]
vault = ["hvac>=2.0"]
gcp = ["google-cloud-secret-manager>=2.16"]
azure = ["azure-keyvault-secrets>=4.7", "azure-identity>=1.15"]
all = [
    "enveloper[aws]",
    "enveloper[vault]",
    "enveloper[gcp]",
    "enveloper[azure]",
    "enveloper[alibaba]",
]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.14",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.11",
    "types-PyYAML>=6.0",
    "twine>=6.0",
]
alibaba = [
    "alibabacloud-kms20160120>=2.0",
]

[project.urls]
Homepage = "https://github.com/raminf/enveloper"
Repository = "https://github.com/raminf/enveloper"
Issues = "https://github.com/raminf/enveloper/issues"

[project.scripts]
enveloper = "enveloper.__main__:main"
envr = "enveloper.__main__:main"

[project.entry-points."enveloper.stores"]
keychain = "enveloper.stores.keychain:KeychainStore"
aws = "enveloper.stores.aws_ssm:AwsSsmStore"
github = "enveloper.stores.github:GitHubStore"
vault = "enveloper.stores.vault:VaultStore"
gcp = "enveloper.stores.gcp_sm:GcpSmStore"
azure = "enveloper.stores.azure_kv:AzureKvStore"
aliyun = "enveloper.stores.aliyun_sm:AliyunSmStore"

[tool.ruff]
target-version = "py310"
line-length = 130

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["enveloper.stores.*", "enveloper.store"]
ignore_missing_imports = true
disable_error_code = ["no-any-return"]

[[tool.mypy.overrides]]
module = "enveloper.resolve_store"
disable_error_code = ["call-arg"]

[[tool.mypy.overrides]]
module = ["enveloper.cli", "enveloper.cli.*"]
disable_error_code = ["attr-defined"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration_aws: integration tests for AWS SSM (set ENVELOPER_TEST_AWS=1 and AWS credentials)",
    "integration_gcp: integration tests for GCP Secret Manager (set ENVELOPER_TEST_GCP=1, ENVELOPER_TEST_GCP_PROJECT, and GCP credentials)",
    "integration_azure: integration tests for Azure Key Vault (set ENVELOPER_TEST_AZURE=1, ENVELOPER_TEST_AZURE_VAULT_URL, and Azure credentials)",
    "integration_alibaba: integration tests for Alibaba Cloud KMS Secrets (set ENVELOPER_TEST_ALIBABA=1, ENVELOPER_TEST_ALIBABA_REGION, and Alibaba credentials)",
]

[tool.uv]
index-url = "https://pypi.org/simple"
